Define service Salesforce
