--- handlers/single-contribution-salesforce-writes/src/main/scala/com/gu/singleContributionSalesforceWrites/services/salesforce/GetApiDomain.scala
@@ -1,10 +1,0 @@
-package com.gu.singleContributionSalesforceWrites.services.salesforce
-
-object GetApiDomain {
-  def apply(stage: String) = {
-    stage match {
-      case "PROD" => "https://gnmtouchpoint.my.salesforce.com"
-      case _ => "https://gnmtouchpoint--dev1.sandbox.my.salesforce.com"
-    }
-  }
-}
